refactor: split transaction file in rmn_btc
