+++ core/src/prelude.rs
@@ -1,8 +1,4 @@
 //! Re-exports of common traits.
 pub use crate::{
-    builder::TxBuilder,
-    enc::AddressEncoder,
-    nets::Network,
-    ser::ByteFormat,
-    types::Transaction,
+    builder::TxBuilder, enc::AddressEncoder, nets::Network, ser::ByteFormat, types::Transaction,
 };
